+++ rl10/design/resources/tss.rb
@@ -2,31 +2,21 @@ module Resources
   class TSS
     include Praxis::ResourceDefinition
 
-    description 'Manipulate the TSS proxy'
+    description 'Manipulate the TSS proxy (this is deprecated, please use the /rll/tss/control resource)'
     media_type 'text/plain'
 
     routing do
       prefix '/rll/tss'
     end
 
-    action :put_control do
-      description 'Control the TSS monitoring'
-      routing { put '/control' }
-      params do
-        attribute :tss_id, String, required: false
-        attribute :enable_monitoring, Attributor::Boolean, required: false
-      end
-      response :ok
-    end
-
     action :get_hostname do
-      description 'Get the TSS hostname to proxy'
+      description 'Get the TSS hostname to proxy (deprecated, RL10 knows the hostname)'
       routing { get '/hostname' }
       response :ok
     end
 
     action :put_hostname do
-      description 'Set the TSS hostname to proxy'
+      description 'Set the TSS hostname to proxy (deprecated, RL10 knows the hostname)'
       routing { put '/hostname' }
       params do
         attribute :hostname, String, required: true
@@ -34,4 +24,44 @@ module Resources
       response :ok
     end
   end
+
+  class TSSControl
+    include Praxis::ResourceDefinition
+
+    description 'Manipulate monitoring (TSS) settings'
+    media_type 'text/plain'
+
+    routing do
+      prefix '/rll/tss/control'
+    end
+
+    action :show do
+      description 'Show monitoring features'
+      routing { get '' }
+      response :ok
+    end
+
+    action :update do
+      description 'Enable/disable monitoring features'
+      routing { put '' }
+      params do
+        attribute :tss_id, String, required: false
+        attribute :enable_monitoring, Attributor::String, required: false,
+		values:['false','true','none','util','extra','all']
+      end
+      response :ok
+    end
+
+    action :put_control do
+      description 'Control the TSS monitoring (deprecated, use the update action)'
+      routing { put '' }
+      params do
+        attribute :tss_id, String, required: false
+        attribute :enable_monitoring, Attributor::String, required: false, example: 'all'
+      end
+      response :ok
+    end
+
+  end
+
 end